backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
requests==2.31.0
beautifulsoup4==4.12.3
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
google-generativeai==0.3.2
python-dotenv==1.0.0
aiohttp==3.9.3
lxml==5.1.0
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
passlib[bcrypt]
python-jose[cryptography]
reportlab==4.4.6
bcrypt==3.2.2
python-multipart==0.0.7
